+++ app/Http/Controllers/api/v1/PermissionController.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Http\Controllers\api\v1;
+
+use App\Http\Controllers\Controller;
+use App\Http\Resources\Permission as PermissionResource;
+use App\Permission;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+
+class PermissionController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return AnonymousResourceCollection
+     */
+    public function index()
+    {
+        return PermissionResource::collection(Permission::all());
+    }
+}