+++ database/migrations/2020_09_25_103057_add_automatic_to_role_user_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddAutomaticToRoleUserTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('role_user', function (Blueprint $table) {
+            $table->boolean('automatic')->default(false);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('role_user', function (Blueprint $table) {
+            $table->dropColumn('automatic');
+        });
+    }
+}