+++ resources/js/app.js
@@ -7,6 +7,7 @@ import router from './router';
 import i18n from './i18n';
 import FlashMessage from '@smartweb/vue-flash-message';
 import Clipboard from 'v-clipboard';
+import env from './env';
 
 Vue.use(Clipboard);
 // Install BootstrapVue
@@ -39,13 +40,13 @@ Vue.config.errorHandler = function (error, vm, info) {
   const responseStatus = error.response !== undefined ? error.response.status : undefined;
   const errorMessage = error.response.data ? error.response.data.message : undefined;
 
-  if (responseStatus === 401) { // 401 => unauthorized, redirect and show error messages as flash!
+  if (responseStatus === env.HTTP_UNAUTHORIZED) { // 401 => unauthorized, redirect and show error messages as flash!
     if (vm.$store.getters['session/isAuthenticated']) {
       vm.flashMessage.info(vm.$t('app.flash.unauthenticated'));
       vm.$store.commit('session/setCurrentUser', null);
       vm.$router.replace({ name: 'login' });
     }
-  } else if (responseStatus === 403 && errorMessage === 'This action is unauthorized.') { // 403 => unauthorized, show error messages as flash!
+  } else if (responseStatus === env.HTTP_FORBIDDEN && errorMessage === 'This action is unauthorized.') { // 403 => unauthorized, show error messages as flash!
     vm.flashMessage.error(vm.$t('app.flash.unauthorized'));
   } else if (responseStatus === 420) { // 420 => only for guests, redirect to home route
     vm.flashMessage.info(vm.$t('app.flash.guestsOnly'));