+++ resources/js/lang/de/settings.js
@@ -2,6 +2,37 @@ export default {
   title: 'Einstellungen',
   description: 'Hier können die Einstellungen der Anwendung verwaltet werden.<br>' +
     'Bitte wählen Sie einen der nebenstehenden Menüpunkte, um die Einstellungen anzupassen.',
+
+  roles: {
+    title: 'Rollen',
+
+    new: 'Neue Rolle erstellen',
+    view: 'Detaillierte Informationen für die Rolle {name}',
+    edit: 'Rolle {name} bearbeiten',
+
+    id: 'ID',
+    name: 'Name',
+    permissions: 'Rechte',
+    roomLimit: {
+      label: 'Max. Anzahl an Räumen',
+      default: 'Systemstandard ({value})',
+      unlimited: 'Unbegrenzt',
+      custom: 'Benutzerdefinierter Wert'
+    },
+
+    default: 'Standard',
+    actions: 'Aktionen',
+    nodata: 'Es sind keine Rollen vorhanden!',
+
+    noOptions: 'Keine Berechtigungen vorhanden!',
+
+    delete: {
+      item: 'Rolle {id} löschen',
+      confirm: 'Wollen Sie die Rolle {name} wirklich löschen?',
+      title: 'Rolle löschen?'
+    }
+  },
+
   users: {
     title: 'Benutzer',
     titleDatabaseCard: 'Datenbank Informationen',
@@ -72,21 +103,5 @@ export default {
       message: 'Etwas ist schief gelaufen! Versuchen sie es erneut',
       title: 'Einladung'
     }
-  },
-
-  rooms: {
-    title: 'Serverräume'
-  },
-
-  recordings: {
-    title: 'Serveraufzeichnungen'
-  },
-
-  siteSettings: {
-    title: 'Seiteneinstellungen'
-  },
-
-  roles: {
-    title: 'Rollen'
   }
 };